sstables: make sstable deletion safer
When system is booting, sstables that don't belong to current shard
are marked for deletion (which actually means agree on deletion of).
That's okay *now* because all sstables loaded during boot have its
shared field set to true, and a shared sstable is only deleted if
all shards agree on its deletion. A shard which a sstable belongs
to will not agree on its deletion.

An unshared sstable, on the other hand, is deleted if marked for
deletion because it doesn't belong to any other shard.
So if in the future we call set_unshared() for an unshared sstable
before the condition for marking sstable for deletion, unshared
sstables will be incorrectly deleted.
Both sstable::set_unshared() and sstable::is_shared() don't say
a word about it.

To make sstable deletion safer, we'll only proceed to deletion
manager on sstable marked for deletion if it's shared *or* it's
unshared and belongs to current shard.
I also added an assert that get_shards_for_this_sstable() should
never return a empty set, or else, all shards would agree on
deletion of the sstable. That could happen if either sharding
metadata or summary metadata is corrupted.

+++ sstables/sstables.cc
@@ -2413,6 +2413,19 @@ sstable::~sstable() {
         // log and ignore this failure, because on startup we'll again try to
         // clean up unused sstables, and because we'll never reuse the same
         // generation number anyway.
+
+        auto belongs_to_current_shard = [this] () -> bool {
+            auto shards = get_shards_for_this_sstable();
+            return boost::range::find(shards, engine().cpu_id()) != shards.end();
+        };
+
+        // a sstable that was set unshared should never be deleted by a shard
+        // which it doesn't belong to.
+        if (!_shared && !belongs_to_current_shard()) {
+            sstlog.warn("Detected attempt to delete unshared sstable {} in a shard that it doesn't belong to", get_filename());
+            return;
+        }
+
         try {
             delete_atomically({sstable_to_delete(filename(component_type::TOC), _shared)}).handle_exception(
                         [op = background_jobs().start()] (std::exception_ptr eptr) {
@@ -2683,6 +2696,9 @@ sstable::get_shards_for_this_sstable() const {
     while (rpras) {
         shards.insert(rpras->shard);
         rpras = sharder.next(*_schema);
+    }
+    if (shards.empty()) {
+        throw std::runtime_error(sprint("Unable to get shards for sstable {}", get_filename()));
     }
     return boost::copy_range<std::vector<unsigned>>(shards);
 }
